Replace Google's libraries-bom with explicit dependencies
The BOM is updated constantly, often without meaningful updates for us.
Also, the Dependabot changelog is often cluttered, which makes it
difficult to check for relevant changes.

As we are only using a single runtime dependency, i. e. the Google Cloud
Storage client, versions are now managed manually. This comes with a
certain probability of incompatible versions of the client and the NIO
library used for testing, which should be acceptable.